refactor(enforcement): rename review-state to block CLI bypass\n\nAssisted-by: PI (claude-opus-4-6-1m) <noreply@pi.dev>
